Revert "dcp-707 Move downloading of geo spreadsheet to geo service"
This reverts commit e4c9b44067b1b0fb526b0ff999b992d371fa9713.
